refactor ainode (has bug needed to be fixed)
