+++ iotdb-core/ainode/iotdb/ainode/manager/model_manager.py
@@ -18,10 +18,11 @@
 from yaml import YAMLError
 
 from iotdb.ainode.constant import TSStatusCode, BuiltInModelType
-from iotdb.ainode.exception import InvaildUriError, BadConfigValueError, BuiltInModelNotSupportError
-from iotdb.ainode.log import logger
+from iotdb.ainode.exception import InvalidUriError, BadConfigValueError, BuiltInModelNotSupportError
+from iotdb.ainode.log import Logger
+from iotdb.ainode.model.built_in_model_factory import fetch_built_in_model
 from iotdb.ainode.model.model_storage import ModelStorage
-from iotdb.ainode.util import get_status
+from iotdb.ainode.util.status import get_status
 from iotdb.thrift.ainode.ttypes import TRegisterModelReq, TRegisterModelResp, TDeleteModelReq
 
 
@@ -30,20 +31,20 @@ class ModelManager:
         self.model_storage = ModelStorage()
 
     def register_model(self, req: TRegisterModelReq):
-        logger.info(f"register model {req.modelId} from {req.uri}")
+        Logger().info(f"register model {req.modelId} from {req.uri}")
         try:
             configs, attributes = self.model_storage.register_model(req.modelId, req.uri)
             return TRegisterModelResp(get_status(TSStatusCode.SUCCESS_STATUS), configs, attributes)
-        except InvaildUriError as e:
-            logger.warning(e)
+        except InvalidUriError as e:
+            Logger().warning(e)
             self.model_storage.delete_model(req.modelId)
             return TRegisterModelResp(get_status(TSStatusCode.INVALID_URI_ERROR, e.message))
         except BadConfigValueError as e:
-            logger.warning(e)
+            Logger().warning(e)
             self.model_storage.delete_model(req.modelId)
             return TRegisterModelResp(get_status(TSStatusCode.INVALID_INFERENCE_CONFIG, e.message))
         except YAMLError as e:
-            logger.warning(e)
+            Logger().warning(e)
             self.model_storage.delete_model(req.modelId)
             if hasattr(e, 'problem_mark'):
                 mark = e.problem_mark
@@ -53,25 +54,26 @@ class ModelManager:
             return TRegisterModelResp(
                 get_status(TSStatusCode.INVALID_INFERENCE_CONFIG, f"An error occurred while parsing the yaml file"))
         except Exception as e:
-            logger.warning(e)
+            Logger().warning(e)
             self.model_storage.delete_model(req.modelId)
             return TRegisterModelResp(get_status(TSStatusCode.AINODE_INTERNAL_ERROR))
 
     def delete_model(self, req: TDeleteModelReq):
-        logger.info(f"delete model {req.modelId}")
+        Logger().info(f"delete model {req.modelId}")
         try:
             self.model_storage.delete_model(req.modelId)
             return get_status(TSStatusCode.SUCCESS_STATUS)
         except Exception as e:
-            logger.warning(e)
+            Logger().warning(e)
             return get_status(TSStatusCode.AINODE_INTERNAL_ERROR, str(e))
 
-
     def load_model(self, model_id: str, acceleration: bool = False):
+        Logger().info(f"load model {model_id}")
+        return self.model_storage.load_model(model_id, acceleration)
+
+    @staticmethod
+    def load_built_in_model(model_id: str, attributes: {}):
         model_id = model_id.lower()
         if model_id not in BuiltInModelType.values():
             raise BuiltInModelNotSupportError(model_id)
-        logger.info(f"load model {model_id}")
-        return self.model_storage.load_model(model_id, acceleration)
-
-
+        return fetch_built_in_model(model_id, attributes)
