+++ iotdb-core/ainode/iotdb/ainode/model/built_in_model_factory.py
@@ -18,9 +18,86 @@
 from abc import abstractmethod
 from typing import List, Dict
 
+import numpy as np
+from sklearn.preprocessing import MinMaxScaler
+from sktime.annotation.hmm_learn import GaussianHMM, GMMHMM
+from sktime.annotation.stray import STRAY
+from sktime.forecasting.arima import ARIMA
+from sktime.forecasting.exp_smoothing import ExponentialSmoothing
+from sktime.forecasting.naive import NaiveForecaster
+from sktime.forecasting.trend import STLForecaster
+
 from iotdb.ainode.constant import AttributeName, BuiltInModelType
+from iotdb.ainode.exception import InferenceModelInternalError, AttributeNotSupportError
 from iotdb.ainode.exception import WrongAttributeTypeError, NumericalRangeException, StringRangeException, \
     ListRangeException, BuiltInModelNotSupportError
+from iotdb.ainode.log import Logger
+
+
+def get_model_attributes(model_id: str):
+    if model_id == BuiltInModelType.ARIMA.value:
+        attribute_map = arima_attribute_map
+    elif model_id == BuiltInModelType.NAIVE_FORECASTER.value:
+        attribute_map = naive_forecaster_attribute_map
+    elif model_id == BuiltInModelType.EXPONENTIAL_SMOOTHING.value:
+        attribute_map = exponential_smoothing_attribute_map
+    elif model_id == BuiltInModelType.STL_FORECASTER.value:
+        attribute_map = stl_forecaster_attribute_map
+    elif model_id == BuiltInModelType.GMM_HMM.value:
+        attribute_map = gmmhmm_attribute_map
+    elif model_id == BuiltInModelType.GAUSSIAN_HMM.value:
+        attribute_map = gaussian_hmm_attribute_map
+    elif model_id == BuiltInModelType.STRAY.value:
+        attribute_map = stray_attribute_map
+    else:
+        raise BuiltInModelNotSupportError(model_id)
+    return attribute_map
+
+
+def fetch_built_in_model(model_id, inference_attributes):
+    """
+    Args:
+        model_id: the unique id of the model
+        inference_attributes: a list of attributes to be inferred, in this function, the attributes will include some
+            parameters of the built-in model. Some parameters are optional, and if the parameters are not
+            specified, the default value will be used.
+    Returns:
+        model: the built-in model
+        attributes: a dict of attributes, where the key is the attribute name, the value is the parsed value of the
+            attribute
+    Description:
+        the create_built_in_model function will create the built-in model, which does not require user
+        registration. This module will parse the inference attributes and create the built-in model.
+    """
+    attribute_map = get_model_attributes(model_id)
+
+    # validate the inference attributes
+    for attribute_name in inference_attributes:
+        if attribute_name not in attribute_map:
+            raise AttributeNotSupportError(model_id, attribute_name)
+
+    # parse the inference attributes, attributes is a Dict[str, Any]
+    attributes = parse_attribute(inference_attributes, attribute_map)
+
+    # build the built-in model
+    if model_id == BuiltInModelType.ARIMA.value:
+        model = ArimaModel(attributes)
+    elif model_id == BuiltInModelType.EXPONENTIAL_SMOOTHING.value:
+        model = ExponentialSmoothingModel(attributes)
+    elif model_id == BuiltInModelType.NAIVE_FORECASTER.value:
+        model = NaiveForecasterModel(attributes)
+    elif model_id == BuiltInModelType.STL_FORECASTER.value:
+        model = STLForecasterModel(attributes)
+    elif model_id == BuiltInModelType.GMM_HMM.value:
+        model = GMMHMMModel(attributes)
+    elif model_id == BuiltInModelType.GAUSSIAN_HMM.value:
+        model = GaussianHmmModel(attributes)
+    elif model_id == BuiltInModelType.STRAY.value:
+        model = STRAYModel(attributes)
+    else:
+        raise BuiltInModelNotSupportError(model_id)
+
+    return model, attributes
 
 
 class Attribute(object):
@@ -29,7 +106,7 @@ class Attribute(object):
         Args:
             name: the name of the attribute
         """
-        self.name = name
+        self._name = name
 
     @abstractmethod
     def get_default_value(self):
@@ -67,7 +144,7 @@ class IntAttribute(Attribute):
         try:
             int_value = int(string_value)
         except:
-            raise WrongAttributeTypeError(self.name, "int")
+            raise WrongAttributeTypeError(self._name, "int")
         return int_value
 
 
@@ -94,7 +171,7 @@ class FloatAttribute(Attribute):
         try:
             float_value = float(string_value)
         except:
-            raise WrongAttributeTypeError(self.name, "float")
+            raise WrongAttributeTypeError(self._name, "float")
         return float_value
 
 
@@ -135,7 +212,7 @@ class BooleanAttribute(Attribute):
         elif string_value.lower() == "false":
             return False
         else:
-            raise WrongAttributeTypeError(self.name, "bool")
+            raise WrongAttributeTypeError(self._name, "bool")
 
 
 class ListAttribute(Attribute):
@@ -163,14 +240,14 @@ class ListAttribute(Attribute):
         try:
             list_value = eval(string_value)
         except:
-            raise WrongAttributeTypeError(self.name, "list")
+            raise WrongAttributeTypeError(self._name, "list")
         if not isinstance(list_value, list):
-            raise WrongAttributeTypeError(self.name, "list")
+            raise WrongAttributeTypeError(self._name, "list")
         for i in range(len(list_value)):
             try:
                 list_value[i] = self.__value_type(list_value[i])
             except:
-                raise ListRangeException(self.name, list_value, self.__type_to_str[self.__value_type])
+                raise ListRangeException(self._name, list_value, self.__type_to_str[self.__value_type])
         return list_value
 
 
@@ -199,15 +276,15 @@ class TupleAttribute(Attribute):
         try:
             tuple_value = eval(string_value)
         except:
-            raise WrongAttributeTypeError(self.name, "tuple")
+            raise WrongAttributeTypeError(self._name, "tuple")
         if not isinstance(tuple_value, tuple):
-            raise WrongAttributeTypeError(self.name, "tuple")
+            raise WrongAttributeTypeError(self._name, "tuple")
         list_value = list(tuple_value)
         for i in range(len(list_value)):
             try:
                 list_value[i] = self.__value_type(list_value[i])
             except:
-                raise ListRangeException(self.name, list_value, self.__type_to_str[self.__value_type])
+                raise ListRangeException(self._name, list_value, self.__type_to_str[self.__value_type])
         tuple_value = tuple(list_value)
         return tuple_value
 
@@ -234,8 +311,9 @@ def parse_attribute(input_attributes: Dict[str, str], attribute_map: Dict[str, A
         else:
             try:
                 attributes[attribute_name] = attribute_map[attribute_name].get_default_value()
-            except:
-                print("attribute_name: ", attribute_name)
+            except NotImplementedError as e:
+                Logger().error(f"attribute {attribute_name} is not implemented.")
+                raise e
     return attributes
 
 
@@ -517,6 +595,7 @@ gaussian_hmm_attribute_map = {
         value_choices=["log", "scaling"],
     )
 }
+
 # GMMHMM
 gmmhmm_attribute_map = {
     AttributeName.N_COMPONENTS.value: IntAttribute(
@@ -649,21 +728,195 @@ stray_attribute_map = {
 }
 
 
-def get_model_attributes(model_id: str):
-    if model_id == BuiltInModelType.ARIMA.value:
-        attribute_map = arima_attribute_map
-    elif model_id == BuiltInModelType.NAIVE_FORECASTER.value:
-        attribute_map = naive_forecaster_attribute_map
-    elif model_id == BuiltInModelType.EXPONENTIAL_SMOOTHING.value:
-        attribute_map = exponential_smoothing_attribute_map
-    elif model_id == BuiltInModelType.STL_FORECASTER.value:
-        attribute_map = stl_forecaster_attribute_map
-    elif model_id == BuiltInModelType.GMM_HMM.value:
-        attribute_map = gmmhmm_attribute_map
-    elif model_id == BuiltInModelType.GAUSSIAN_HMM.value:
-        attribute_map = gaussian_hmm_attribute_map
-    elif model_id == BuiltInModelType.STRAY.value:
-        attribute_map = stray_attribute_map
-    else:
-        raise BuiltInModelNotSupportError(model_id)
-    return attribute_map
+class BuiltInModel(object):
+    def __init__(self, attributes):
+        self._attributes = attributes
+        self._model = None
+
+    @abstractmethod
+    def inference(self, data):
+        raise NotImplementedError
+
+
+class ArimaModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(ArimaModel, self).__init__(attributes)
+        self._model = ARIMA(
+            order=attributes['order'],
+            seasonal_order=attributes['seasonal_order'],
+            method=attributes['method'],
+            suppress_warnings=attributes['suppress_warnings'],
+            maxiter=attributes['maxiter'],
+            out_of_sample_size=attributes['out_of_sample_size'],
+            scoring=attributes['scoring'],
+            with_intercept=attributes['with_intercept'],
+            time_varying_regression=attributes['time_varying_regression'],
+            enforce_stationarity=attributes['enforce_stationarity'],
+            enforce_invertibility=attributes['enforce_invertibility'],
+            simple_differencing=attributes['simple_differencing'],
+            measurement_error=attributes['measurement_error'],
+            mle_regression=attributes['mle_regression'],
+            hamilton_representation=attributes['hamilton_representation'],
+            concentrate_scale=attributes['concentrate_scale']
+        )
+
+    def inference(self, data):
+        try:
+            predict_length = self._attributes['predict_length']
+            self._model.fit(data)
+            output = self._model.predict(fh=range(predict_length))
+            output = np.array(output, dtype=np.float64)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
+
+
+class ExponentialSmoothingModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(ExponentialSmoothingModel, self).__init__(attributes)
+        self._model = ExponentialSmoothing(
+            damped_trend=attributes['damped_trend'],
+            initialization_method=attributes['initialization_method'],
+            optimized=attributes['optimized'],
+            remove_bias=attributes['remove_bias'],
+            use_brute=attributes['use_brute']
+        )
+
+    def inference(self, data):
+        try:
+            predict_length = self._attributes['predict_length']
+            self._model.fit(data)
+            output = self._model.predict(fh=range(predict_length))
+            output = np.array(output, dtype=np.float64)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
+
+
+class NaiveForecasterModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(NaiveForecasterModel, self).__init__(attributes)
+        self._model = NaiveForecaster(
+            strategy=attributes['strategy'],
+            sp=attributes['sp']
+        )
+
+    def inference(self, data):
+        try:
+            predict_length = self._attributes['predict_length']
+            self._model.fit(data)
+            output = self._model.predict(fh=range(predict_length))
+            output = np.array(output, dtype=np.float64)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
+
+
+class STLForecasterModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(STLForecasterModel, self).__init__(attributes)
+        self._model = STLForecaster(
+            sp=attributes['sp'],
+            seasonal=attributes['seasonal'],
+            seasonal_deg=attributes['seasonal_deg'],
+            trend_deg=attributes['trend_deg'],
+            low_pass_deg=attributes['low_pass_deg'],
+            seasonal_jump=attributes['seasonal_jump'],
+            trend_jump=attributes['trend_jump'],
+            low_pass_jump=attributes['low_pass_jump']
+        )
+
+    def inference(self, data):
+        try:
+            predict_length = self._attributes['predict_length']
+            self._model.fit(data)
+            output = self._model.predict(fh=range(predict_length))
+            output = np.array(output, dtype=np.float64)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
+
+
+class GMMHMMModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(GMMHMMModel, self).__init__(attributes)
+        self._model = GMMHMM(
+            n_components=attributes['n_components'],
+            n_mix=attributes['n_mix'],
+            min_covar=attributes['min_covar'],
+            startprob_prior=attributes['startprob_prior'],
+            transmat_prior=attributes['transmat_prior'],
+            means_prior=attributes['means_prior'],
+            means_weight=attributes['means_weight'],
+            weights_prior=attributes['weights_prior'],
+            algorithm=attributes['algorithm'],
+            covariance_type=attributes['covariance_type'],
+            n_iter=attributes['n_iter'],
+            tol=attributes['tol'],
+            params=attributes['params'],
+            init_params=attributes['init_params'],
+            implementation=attributes['implementation']
+        )
+
+    def inference(self, data):
+        try:
+            self._model.fit(data)
+            output = self._model.predict(data)
+            output = np.array(output, dtype=np.int32)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
+
+
+class GaussianHmmModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(GaussianHmmModel, self).__init__(attributes)
+        self._model = GaussianHMM(
+            n_components=attributes['n_components'],
+            covariance_type=attributes['covariance_type'],
+            min_covar=attributes['min_covar'],
+            startprob_prior=attributes['startprob_prior'],
+            transmat_prior=attributes['transmat_prior'],
+            means_prior=attributes['means_prior'],
+            means_weight=attributes['means_weight'],
+            covars_prior=attributes['covars_prior'],
+            covars_weight=attributes['covars_weight'],
+            algorithm=attributes['algorithm'],
+            n_iter=attributes['n_iter'],
+            tol=attributes['tol'],
+            params=attributes['params'],
+            init_params=attributes['init_params'],
+            implementation=attributes['implementation']
+        )
+
+    def inference(self, data):
+        try:
+            self._model.fit(data)
+            output = self._model.predict(data)
+            output = np.array(output, dtype=np.int32)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
+
+
+class STRAYModel(BuiltInModel):
+    def __init__(self, attributes):
+        super(STRAYModel, self).__init__(attributes)
+        self._model = STRAY(
+            alpha=attributes['alpha'],
+            k=attributes['k'],
+            knn_algorithm=attributes['knn_algorithm'],
+            p=attributes['p'],
+            size_threshold=attributes['size_threshold'],
+            outlier_tail=attributes['outlier_tail']
+        )
+
+    def inference(self, data):
+        try:
+            data = MinMaxScaler().fit_transform(data)
+            output = self._model.fit_transform(data)
+            # change the output to int
+            output = np.array(output, dtype=np.int32)
+            return output
+        except Exception as e:
+            raise InferenceModelInternalError(str(e))
