+++ iotdb-core/ainode/iotdb/ainode/model/model_factory.py
@@ -25,9 +25,10 @@ from requests.adapters import HTTPAdapter
 
 from iotdb.ainode.constant import DEFAULT_RECONNECT_TIMES, DEFAULT_RECONNECT_TIMEOUT, DEFAULT_CHUNK_SIZE, \
     DEFAULT_CONFIG_FILE_NAME, DEFAULT_MODEL_FILE_NAME
-from iotdb.ainode.exception import InvaildUriError
-from iotdb.ainode.log import logger
-from iotdb.ainode.parser import parse_inference_config
+from iotdb.ainode.exception import InvalidUriError, BadConfigValueError
+from iotdb.ainode.log import Logger
+from iotdb.ainode.util.serde import get_data_type_byte_from_str
+from iotdb.thrift.ainode.ttypes import TConfigs
 
 HTTP_PREFIX = "http://"
 HTTPS_PREFIX = "https://"
@@ -66,7 +67,7 @@ def _download_file(url: str, storage_path: str) -> None:
     Returns:
         None
     """
-    logger.debug(f"download file from {url} to {storage_path}")
+    Logger().debug(f"download file from {url} to {storage_path}")
 
     session = Session()
     adapter = HTTPAdapter(max_retries=DEFAULT_RECONNECT_TIMES)
@@ -81,7 +82,7 @@ def _download_file(url: str, storage_path: str) -> None:
             if chunk:
                 file.write(chunk)
 
-    logger.debug(f"download file from {url} to {storage_path} success")
+    Logger().debug(f"download file from {url} to {storage_path} success")
 
 
 def _register_model_from_network(uri: str, model_storage_path: str,
@@ -107,7 +108,7 @@ def _register_model_from_network(uri: str, model_storage_path: str,
     # read and parse config dict from config.yaml
     with open(config_storage_path, 'r', encoding='utf-8') as file:
         config_dict = yaml.safe_load(file)
-    configs, attributes = parse_inference_config(config_dict)
+    configs, attributes = _parse_inference_config(config_dict)
 
     # if config.yaml is correct, download model file
     _download_file(target_model_path, model_storage_path)
@@ -138,24 +139,91 @@ def _register_model_from_local(uri: str, model_storage_path: str,
     attributes = None
     if exist_model_file and exist_config_file:
         # copy config.yaml
-        logger.debug(f"copy file from {target_config_path} to {config_storage_path}")
+        Logger().debug(f"copy file from {target_config_path} to {config_storage_path}")
         shutil.copy(target_config_path, config_storage_path)
-        logger.debug(f"copy file from {target_config_path} to {config_storage_path} success")
+        Logger().debug(f"copy file from {target_config_path} to {config_storage_path} success")
 
         # read and parse config dict from config.yaml
         with open(config_storage_path, 'r', encoding='utf-8') as file:
             config_dict = yaml.safe_load(file)
-        configs, attributes = parse_inference_config(config_dict)
+        configs, attributes = _parse_inference_config(config_dict)
 
         # if config.yaml is correct, copy model file
-        logger.debug(f"copy file from {target_model_path} to {model_storage_path}")
+        Logger().debug(f"copy file from {target_model_path} to {model_storage_path}")
         shutil.copy(target_model_path, model_storage_path)
-        logger.debug(f"copy file from {target_model_path} to {model_storage_path} success")
+        Logger().debug(f"copy file from {target_model_path} to {model_storage_path} success")
 
     elif not exist_model_file or not exist_config_file:
-        raise InvaildUriError(uri)
+        raise InvalidUriError(uri)
 
     return configs, attributes
+
+
+def _parse_inference_config(config_dict):
+    """
+    Args:
+        config_dict: dict
+            - configs: dict
+                - input_shape (list<i32>): input shape of the model and needs to be two-dimensional array like [96, 2]
+                - output_shape (list<i32>): output shape of the model and needs to be two-dimensional array like [96, 2]
+                - input_type (list<str>): input type of the model and each element needs to be in ['bool', 'int32', 'int64', 'float32', 'float64', 'text'], default float64
+                - output_type (list<str>): output type of the model and each element needs to be in ['bool', 'int32', 'int64', 'float32', 'float64', 'text'], default float64
+            - attributes: dict
+    Returns:
+        configs: TConfigs
+        attributes: str
+    """
+    configs = config_dict['configs']
+
+    # check if input_shape and output_shape are two-dimensional array
+    if not (isinstance(configs['input_shape'], list) and len(configs['input_shape']) == 2):
+        raise BadConfigValueError('input_shape', configs['input_shape'],
+                                  'input_shape should be a two-dimensional array.')
+    if not (isinstance(configs['output_shape'], list) and len(configs['output_shape']) == 2):
+        raise BadConfigValueError('output_shape', configs['output_shape'],
+                                  'output_shape should be a two-dimensional array.')
+
+    # check if input_shape and output_shape are positive integer
+    input_shape_is_positive_number = isinstance(configs['input_shape'][0], int) and isinstance(
+        configs['input_shape'][1], int) and configs['input_shape'][0] > 0 and configs['input_shape'][1] > 0
+    if not input_shape_is_positive_number:
+        raise BadConfigValueError('input_shape', configs['input_shape'],
+                                  'element in input_shape should be positive integer.')
+
+    output_shape_is_positive_number = isinstance(configs['output_shape'][0], int) and isinstance(
+        configs['output_shape'][1], int) and configs['output_shape'][0] > 0 and configs['output_shape'][1] > 0
+    if not output_shape_is_positive_number:
+        raise BadConfigValueError('output_shape', configs['output_shape'],
+                                  'element in output_shape should be positive integer.')
+
+    # check if input_type and output_type are one-dimensional array with right length
+    if 'input_type' in configs and not (
+            isinstance(configs['input_type'], list) and len(configs['input_type']) == configs['input_shape'][1]):
+        raise BadConfigValueError('input_type', configs['input_type'],
+                                  'input_type should be a one-dimensional array and length of it should be equal to input_shape[1].')
+
+    if 'output_type' in configs and not (
+            isinstance(configs['output_type'], list) and len(configs['output_type']) == configs['output_shape'][1]):
+        raise BadConfigValueError('output_type', configs['output_type'],
+                                  'output_type should be a one-dimensional array and length of it should be equal to output_shape[1].')
+
+    # parse input_type and output_type to byte
+    if 'input_type' in configs:
+        input_type = [get_data_type_byte_from_str(x) for x in configs['input_type']]
+    else:
+        input_type = [get_data_type_byte_from_str('float32')] * configs['input_shape'][1]
+
+    if 'output_type' in configs:
+        output_type = [get_data_type_byte_from_str(x) for x in configs['output_type']]
+    else:
+        output_type = [get_data_type_byte_from_str('float32')] * configs['output_shape'][1]
+
+    # parse attributes
+    attributes = ""
+    if 'attributes' in config_dict:
+        attributes = str(config_dict['attributes'])
+
+    return TConfigs(configs['input_shape'], configs['output_shape'], input_type, output_type), attributes
 
 
 def fetch_model_by_uri(uri: str, model_storage_path: str, config_storage_path: str):
