+++ iotdb-core/ainode/iotdb/ainode/util/decorator.py
@@ -15,3 +15,12 @@
 # specific language governing permissions and limitations
 # under the License.
 #
+def singleton(cls):
+    instances = {}
+
+    def get_instance(*args, **kwargs):
+        if cls not in instances:
+            instances[cls] = cls(*args, **kwargs)
+        return instances[cls]
+
+    return get_instance
